moving price component from Price folder to results
