+++ frontend/src/components/Results/FX_RATES_ANNUAL-sd-2017-01-01.json
